refactor: draw-frames.js -> draw-frames.ts
