+++ sources/canvas/renderer.js
@@ -7,7 +7,7 @@ import { getImageToDraw } from "./palette-recolor.js";
 import { getMultiRecolors } from "../state/palettes.js";
 import { get2DContext, getZPos } from "./canvas-utils.js";
 import { variantToFilename } from "../utils/helpers.js";
-import { drawFramesToCustomAnimation } from "./draw-frames.js";
+import { drawFramesToCustomAnimation } from "./draw-frames.ts";
 import {
   FRAME_SIZE,
   ANIMATION_OFFSETS,
